secure request index of CoupleController + secure requests of AccountController + change redirect when user not log in

+++ app-site/src/Controller/DetectionController.php
@@ -27,6 +27,11 @@ final class DetectionController extends AbstractController{
     #[Route('/detections', name: 'app_detections')]
     public function index(UserInterface $user, PaginatorInterface $paginator, Request $request): Response
     {
+        if (!$user) {
+            $this->addFlash('error', 'You need to sign in to see the detections !');
+            return $this->redirectToRoute('app_login');
+        }
+
         $query = $this->detectionService->getAllDetectionsByUser($user->getId());
 
         $detections = $paginator->paginate(
@@ -41,12 +46,22 @@ final class DetectionController extends AbstractController{
     }
 
     #[Route('/detections/view/{id}', name: 'app_detections_view')]
-    public function viewDetection(string $id): Response
+    public function viewDetection(string $id, UserInterface $user): Response
     {
         $detection = $this->detectionService->getDetectionById($id);
         if (!$detection) {
             $this->addFlash('error', 'Detection not found');
             return $this->redirectToRoute('app_detections');
+        }
+
+        // Check authorization
+        if (!$user) {
+            $this->addFlash('error', 'You need to sign in to see this detection !');
+            return $this->redirectToRoute('app_login');
+        }
+        if ($user->getUserIdentifier() != $detection->getCouple()->getUser()->getUsername()) {
+            $this->addFlash('error', 'You are not authorized to see this detection !');
+            return $this->redirectToRoute('app_couples');
         }
 
         return $this->render('detection/view.html.twig', [
@@ -55,12 +70,22 @@ final class DetectionController extends AbstractController{
     }
 
     #[Route('/detections/delete/{id}', name: 'app_detections_delete')]
-    public function deleteDetection(string $id): Response
+    public function deleteDetection(string $id, UserInterface $user): Response
     {
         $detection = $this->detectionService->getDetectionById($id);
         if (!$detection) {
             $this->addFlash('error', 'Detection not found');
             return $this->redirectToRoute('app_detections');
+        }
+
+        // Check authorization
+        if (!$user) {
+            $this->addFlash('error', 'You need to sign in to delete this detection !');
+            return $this->redirectToRoute('app_login');
+        }
+        if ($user->getUserIdentifier() != $detection->getCouple()->getUser()->getUsername()) {
+            $this->addFlash('error', 'You are not authorized to delete this detection !');
+            return $this->redirectToRoute('app_couples');
         }
 
         $this->entityManager->remove($detection);
@@ -73,10 +98,12 @@ final class DetectionController extends AbstractController{
     #[Route('/detections/image/{filename}', name: 'app_detections_image', methods: ['GET'])]
     public function getProtectedImage(string $filename, UserInterface $user): Response
     {
-        // 1. Auth check
+        // Auth check
         if (!$user) {
-            return $this->apiResponseService->error('You are not authorized to access this stream! Sign in first!');
+            $this->addFlash('error', 'You are not authorized to access this stream! Sign in first!');
+            return $this->redirectToRoute('app_login');
         }
+
 
         // read the image file
         $filePath = $this->parameterBag->get('detections_dir') . "/" . $filename;
